add: panels register
